Move original views/serializers to v1 module
