+++ EquiTrack/EquiTrack/urls.py
@@ -36,7 +36,7 @@ from trips.views import TripsViewSet, TripFileViewSet, TripActionPointViewSet
 from partners.views import PartnerOrganizationsViewSet, AgreementViewSet, PartnerStaffMembersViewSet, FileTypeViewSet
 from users.views import UserViewSet, GroupViewSet, OfficeViewSet, SectionViewSet
 from funds.views import DonorViewSet, GrantViewSet
-from reports.views import (
+from reports.views.v1 import (
     ResultStructureViewSet,
     ResultTypeViewSet,
     SectorViewSet,
