+++ EquiTrack/reports/views/v1.py
@@ -2,7 +2,7 @@ __author__ = 'achamseddine'
 
 from rest_framework import viewsets, mixins
 from rest_framework.permissions import IsAdminUser
-from .models import (
+from reports.models import (
     ResultStructure,
     ResultType,
     Result,
@@ -11,7 +11,7 @@ from .models import (
     Indicator,
     Unit
 )
-from .serializers import (
+from reports.serializers.v1 import (
     ResultStructureSerializer,
     ResultTypeSerializer,
     ResultSerializer,
